requirements.txt
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
fonttools==4.44.3
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.23.5
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
-e .